pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "chat-oracle"
version = "0.1.0"
description = "LLM gateway and subscription control plane."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.14.0",
    "asyncpg>=0.30.0",
    "bcrypt>=4.2.0",
    "crawl4ai>=0.8.0",
    "email-validator>=2.2.0",
    "fastapi>=0.115.0",
    "httpx>=0.27.0",
    "orjson>=3.10.0",
    "pydantic-settings>=2.6.0",
    "psycopg2-binary>=2.9.10",
    "python-jose[cryptography]>=3.3.0",
    "python-multipart>=0.0.12",
    "redis>=5.2.0",
    "SQLAlchemy>=2.0.36",
    "structlog>=24.4.0",
    "telethon>=1.43.0",
    "uvicorn[standard]>=0.32.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
]

[tool.setuptools.packages.find]
include = ["app*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
